Improve markdown preview, other job posting UX improvements
Better in-input label plugin, removed a lot of unnecessary JS that used
to do something I'm sure. The preview is now a tab, instead of having to
scroll down. It is also a more accurate representation of the final
product. I even got selfish and improved my admin interface a bit.