Polish script for setting up new developer environments
When setting up a new environment, a developer has to create a
database.yml from scratch in order to complete setup. This adds a
database.yml.sample that can be copied over and makes use of it in the
bin/setup script. The readme was changed to reference the bin/setup
script in the instructions for setting up a new environment.

One other minor paint point was setting up and accessing the admin pages
for the first time. To help with this, I added an admin user to the seed
file and instructions in the readme on how to access the admin pages
with that user.