Do not delete jobs; only show active jobs
Why?
Resolves #81
Resolves #107

We don't want to delete job listing that are no longer active so we have
them for historical reference but we don't want to show them in public
listings after they are no longer active.

What?
Introduces the concept of active job postings. Job listings are shown in
public listings if they are active.