Update Rails to 6.1.3.1
Considering the recent "super happy GPL funtime" event that recently
happened, it may not be the worst idea to go ahead and update Rails to
the latest.

This updates the Rails gem and its support gems to the latest versions.
This includes the removal of mimemagic in place of marcel (which I don't
think the app actually uses, but it's nice to be current).